add packets task_repository and model
