+++ TaskDone.go
@@ -13,7 +13,7 @@ func apiTaskDone(w http.ResponseWriter, r *http.Request) {
 
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 
-	task, err := taskRepo.getbyID(id)
+	task, err := taskRepo.GetbyID(id)
 	if err != nil {
 		jsonError(w, "Задача не найдена", err)
 		return
