+++ addTask.go
@@ -6,10 +6,12 @@ import (
 	"fmt"
 	"net/http"
 	"time"
+
+	"github.com/GalievRinat/go_final_project/model"
 )
 
 func apiAddTask(w http.ResponseWriter, r *http.Request) {
-	var task Task
+	var task model.Task
 	var buf bytes.Buffer
 
 	_, err := buf.ReadFrom(r.Body)
