+++ getTasks.go
@@ -4,12 +4,14 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+
+	"github.com/GalievRinat/go_final_project/model"
 )
 
 func apiGetTasks(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 
-	tasks, err := taskRepo.getAll()
+	tasks, err := taskRepo.GetAll()
 	if err != nil {
 		jsonError(w, "Ошибка получения списка задач", err)
 		return
@@ -17,10 +19,10 @@ func apiGetTasks(w http.ResponseWriter, r *http.Request) {
 
 	fmt.Println(tasks)
 	if tasks == nil {
-		tasks = make([]Task, 0)
+		tasks = make([]model.Task, 0)
 	}
 
-	resp, err := json.Marshal(map[string][]Task{"tasks": tasks})
+	resp, err := json.Marshal(map[string][]model.Task{"tasks": tasks})
 	if err != nil {
 		jsonError(w, "Ошибка сериализации JSON", err)
 		return
